Specify a name property for the jobs and rename gulp task file to fix typo
