+++ .github/workflows/build.yaml
@@ -15,6 +15,7 @@ permissions:
 
 jobs:
     build:
+        name: Build TypeScript Project
         uses: OutSystems/ui-components.github-reusable-workflows/.github/workflows/ts-build-project.yaml@7ffd7f90f6e5016bd92d0e050d2516084b99c2f6 #v0.2.5
         with:
             github-ref: ${{ github.ref }}
