+++ .github/workflows/validate-pr-labels.yaml
@@ -14,4 +14,5 @@ permissions:
 
 jobs:
     check-label:
+        name: Validate PR Labels
         uses: OutSystems/rd.github-reusable-workflows/.github/workflows/validate-pr-labels.yaml@9d497d1c5bc6e355aa8f4663539e6b75c212f6b4 #v2.0.7
