+++ .github/workflows/validate-pr-title.yaml
@@ -14,6 +14,7 @@ permissions:
 
 jobs:
     check-title:
+        name: Validate PR Title
         uses: OutSystems/rd.github-reusable-workflows/.github/workflows/validate-pr-title.yaml@9d497d1c5bc6e355aa8f4663539e6b75c212f6b4 #v2.0.7
         with:
             validate-semVer: false
